Increase test coverage to 50%+ and fix clippy warnings
- Added 1276+ tests across all pure-logic modules to push from 29.2% to 50.02% region coverage
- New test files: combat/input/targeting.rs (inline), enemy/tests.rs (expanded),
  world/tests.rs, world/starmap/tests.rs, world/location_gen/tests.rs,
  combat/terrain/tests.rs, combat/action/tests.rs, combat/ai/tests.rs,
  combat/grid/tests.rs, combat/synergy/tests.rs, particle/tests.rs,
  radical/tests.rs, codex/tests.rs, crucible.rs (inline)
- Comprehensive RadicalAction/PlayerRadicalAbility variant coverage
- Tests for spell_range, compute_aoe_preview, line_between, dash_target_tiles, blink_target_tiles
- Fixed 193 clippy warnings via cargo clippy --fix and manual fixes
- Suppressed unstable is_multiple_of lint with allow attribute in lib.rs
  (is_multiple_of is gated behind #![feature(unsigned_is_multiple_of)] on MSVC)
- Fixed negated modulo expressions introduced by auto-fix
- 0 clippy warnings, 2688 tests, 50.02% region coverage

+++ src/combat/transition.rs
@@ -596,7 +596,7 @@ fn generate_arena(floor: i32, size: usize, biome: ArenaBiome) -> TacticalArena {
     if floor >= 4 {
         let flow_seed = seed.wrapping_mul(6271).wrapping_add(99);
         let flow_hash = flow_seed.wrapping_mul(31337);
-        let should_place = flow_hash % 3 != 0;
+        let should_place = flow_hash % 3 != 0 ;
         if should_place {
             let dir = flow_hash % 4;
             let (flow_tile, is_horizontal) = match dir {
@@ -605,7 +605,7 @@ fn generate_arena(floor: i32, size: usize, biome: ArenaBiome) -> TacticalArena {
                 2 => (BattleTile::ConveyorE, true),
                 _ => (BattleTile::ConveyorW, true),
             };
-            let channel_len = 3 + (floor / 6).min(3) as i32;
+            let channel_len = 3 + (floor / 6).min(3);
             if is_horizontal {
                 let y = (2 + (flow_hash >> 8) % (size as u64 - 4)) as i32;
                 let start_x = ((flow_hash >> 16) % (size as u64 / 2)) as i32;
